requirements.txt
streamlit==1.28.0
groq==0.32.0
pyserial==3.5
plotly==5.17.0
pandas==2.1.3
python-dotenv==1.0.0